(PC-18978) test(OfferDuoModal): fix test for OfferDuoModal
