+++ src/features/search/pages/modals/OfferDuoModal/OfferDuoModal.web.test.tsx
@@ -1,19 +1,19 @@
 import React from 'react'
 
-import { OfferTypeModal } from 'features/search/pages/modals/OfferTypeModal/OfferTypeModal'
+import { OfferDuoModal } from 'features/search/pages/modals/OfferDuoModal/OfferDuoModal'
 import { checkAccessibilityFor, render } from 'tests/utils/web'
 
 jest.mock('react-query')
 
 jest.mock('features/auth/AuthContext')
 
-describe('<OfferTypeModal/>', () => {
+describe('<OfferDuoModal/>', () => {
   describe('Accessibility', () => {
     it('should not have basic accessibility issues', async () => {
       const { container } = render(
-        <OfferTypeModal
-          title="Type d'offre"
-          accessibilityLabel="Ne pas filtrer sur les type d'offre et retourner aux résultats"
+        <OfferDuoModal
+          title="Duo"
+          accessibilityLabel="Ne pas filtrer sur les offres duo et retourner aux résultats"
           isVisible
           hideModal={jest.fn()}
         />
